refactor: move patch from `__init__` to `_compat`
